Combine *_containers_of_type
